+++ Dockerfile
@@ -1,15 +1,26 @@
 ARG BROWSER_VERSION=90
 
-FROM oldwebtoday/chrome:${BROWSER_VERSION} as chrome
+ARG BROWSER_IMAGE_BASE=oldwebtoday/chrome
 
-FROM nikolaik/python-nodejs:python3.8-nodejs14
+FROM ${BROWSER_IMAGE_BASE}:${BROWSER_VERSION} as chrome
 
-RUN curl -sS https://dl.yarnpkg.com/debian/pubkey.gpg | apt-key add -
+#FROM nikolaik/python-nodejs:python3.9-nodejs16
+FROM ubuntu:bionic
 
+#RUN apt-get update -y && apt-get install --no-install-recommends -qqy software-properties-common \
+#&& add-apt-repository -y ppa:deadsnakes \
 RUN apt-get update -y \
-    && apt-get install --no-install-recommends -qqy fonts-stix locales-all redis-server xvfb \
+    && apt-get install --no-install-recommends -qqy build-essential fonts-stix locales-all redis-server xvfb gpg-agent curl git \
+		python3 python3-distutils python3-dev gpg ca-certificates \
     && apt-get clean \
     && rm -rf /var/lib/apt/lists/*
+
+RUN curl -sS https://dl.yarnpkg.com/debian/pubkey.gpg | apt-key add - \
+    && echo "deb https://dl.yarnpkg.com/debian/ stable main" | tee /etc/apt/sources.list.d/yarn.list \
+    && curl -sL https://deb.nodesource.com/setup_16.x -o /tmp/nodesource_setup.sh && bash /tmp/nodesource_setup.sh \
+		&& apt-get update -y && apt-get install -qqy nodejs yarn \
+    && curl https://bootstrap.pypa.io/get-pip.py | python3 \
+    && pip3 install -U setuptools
 
 ARG BROWSER_VERSION
 
@@ -29,7 +40,7 @@ RUN dpkg -i /deb/*.deb; apt-get update; apt-get install -fqqy && \
 WORKDIR /app
 
 ADD requirements.txt /app/
-RUN pip install -r requirements.txt
+RUN pip3 install -r requirements.txt
 
 ADD package.json /app/
