+++ util/argParser.js
@@ -5,36 +5,192 @@ const path = require("path");
 const fs = require("fs");
 const child_process = require("child_process");
 const { NewWindowPage} = require("./screencaster");
-const { constants } = require("./constants");
+const { BEHAVIOR_LOG_FUNC, WAIT_UNTIL_OPTS } = require("./constants");
+const yargs = require("yargs/yargs");
+const { hideBin } = require("yargs/helpers");
 
 // ============================================================================
-class argParser {
-  constructor(){
-    this.constants = new constants();
+class ArgParser {
+  get cliOpts() {
+    return {
+      "url": {
+        alias: "u",
+        describe: "The URL to start crawling from",
+        type: "string",
+      },
+
+      "workers": {
+        alias: "w",
+        describe: "The number of workers to run in parallel",
+        default: 1,
+        type: "number",
+      },
+
+      "newContext": {
+        describe: "The context for each new capture, can be a new: page, window, session or browser.",
+        default: "page",
+        type: "string"
+      },
+
+      "waitUntil": {
+        describe: "Puppeteer page.goto() condition to wait for before continuing, can be multiple separate by ','",
+        default: "load,networkidle0",
+      },
+
+      "limit": {
+        describe: "Limit crawl to this number of pages",
+        default: 0,
+        type: "number",
+      },
+
+      "timeout": {
+        describe: "Timeout for each page to load (in seconds)",
+        default: 90,
+        type: "number",
+      },
+
+      "scope": {
+        describe: "Regex of page URLs that should be included in the crawl (defaults to the immediate directory of URL)",
+      },
+
+      "scopeType": {
+        describe: "Simplified scope for which URLs to crawl, can be: prefix, page, domain, any",
+        type: "string",
+      },
+
+      "exclude": {
+        describe: "Regex of page URLs that should be excluded from the crawl."
+      },
+
+      "allowHashUrls": {
+        describe: "Allow Hashtag URLs, useful for single-page-application crawling or when different hashtags load dynamic content",
+      },
+
+      "collection": {
+        alias: "c",
+        describe: "Collection name to crawl to (replay will be accessible under this name in pywb preview)",
+        type: "string",
+        default: `capture-${new Date().toISOString().slice(0,19)}`.replace(/:/g, "-")
+      },
+
+      "headless": {
+        describe: "Run in headless mode, otherwise start xvfb",
+        type: "boolean",
+        default: false,
+      },
+
+      "driver": {
+        describe: "JS driver for the crawler",
+        type: "string",
+        default: path.join(__dirname, "..", "defaultDriver.js"),
+      },
+
+      "generateCDX": {
+        alias: ["generatecdx", "generateCdx"],
+        describe: "If set, generate index (CDXJ) for use with pywb after crawl is done",
+        type: "boolean",
+        default: false,
+      },
+
+      "combineWARC": {
+        alias: ["combinewarc", "combineWarc"],
+        describe: "If set, combine the warcs",
+        type: "boolean",
+        default: false,
+      },
+      
+      "rolloverSize": {
+        describe: "If set, declare the rollover size",
+        default: 1000000000,
+        type: "number",
+      },
+      
+      "generateWACZ": {
+        alias: ["generatewacz", "generateWacz"],
+        describe: "If set, generate wacz",
+        type: "boolean",
+        default: false,
+      },
+      
+      "logging": {
+        describe: "Logging options for crawler, can include: stats, pywb, behaviors, behaviors-debug",
+        type: "string",
+        default: "stats",
+      },
+      
+      "urlFile": {
+        alias: ["urlfile", "url-file", "url-list"],
+        describe: "If set, read a list of urls from the passed file INSTEAD of the url from the --url flag.",
+        type: "string",
+      },
+      
+      "text": {
+        describe: "If set, extract text to the pages.jsonl file",
+        type: "boolean",
+        default: false,
+      },
+      
+      "cwd": {
+        describe: "Crawl working directory for captures (pywb root). If not set, defaults to process.cwd()",
+        type: "string",
+        default: process.cwd(),
+      },
+
+      "mobileDevice": {
+        describe: "Emulate mobile device by name from: https://github.com/puppeteer/puppeteer/blob/main/src/common/DeviceDescriptors.ts",
+        type: "string",
+      },
+
+      "userAgent": {
+        describe: "Override user-agent with specified string",
+        type: "string",
+      },
+
+      "userAgentSuffix": {
+        describe: "Append suffix to existing browser user-agent (ex: +MyCrawler, info@example.com)",
+        type: "string",
+      },
+
+      "useSitemap": {
+        describe: "If enabled, check for sitemaps at /sitemap.xml, or custom URL if URL is specified",
+      },
+
+      "statsFilename": {
+        describe: "If set, output stats as JSON to this file. (Relative filename resolves to crawl working directory)"
+      },
+
+      "behaviors": {
+        describe: "Which background behaviors to enable on each page",
+        default: "autoplay,autofetch,siteSpecific",
+        type: "string",
+      },
+
+      "profile": {
+        describe: "Path to tar.gz file which will be extracted and used as the browser profile",
+        type: "string",
+      },
+
+      "screencastPort": {
+        describe: "If set to a non-zero value, starts an HTTP server with screencast accessible on this port",
+        type: "number",
+        default: 0
+      },
+    };
   }
 
-  parseYaml(yamlConfigFile){
-    try {
-      console.log("YAML config detected. The values declared in this file will be used and any command line flags passed will override them");
-
-      var fileContents = fs.readFileSync(yamlConfigFile, "utf8");
-
-      var data = yaml.safeLoad(fileContents);
-
-      if (!data.crawler){
-        console.log("Error parsing the yaml file: Yaml config file needs to have the arguments under 'crawler' field please see the github readme for more details on the yaml configuration");
-        return false;
-      }
-
-      this.validateArgs(data.crawler);
-      return data.crawler;
-    }
-    catch (e) {
-      console.log(e);
-      return false;
-    }
+  parseArgs(argv) {
+    argv = argv || process.argv;
+    
+    return yargs(hideBin(argv))
+      .usage("crawler [options]")
+      .option(this.cliOpts)
+      .config("yamlConfig", (configPath) => {
+        return yaml.safeLoad(fs.readFileSync(configPath, "utf-8"));
+      })
+      .check((argv) => this.validateArgs(argv))
+      .argv;
   }
-
+ 
   rxEscape(string) {
     return string.replace(/[-/\\^$*+?.()|[\]{}]/g, "\\$&");
   }
@@ -51,6 +207,16 @@ class argParser {
 
 
   validateArgs(argv) {
+    /*
+    if (argv.yamlConfig) {
+			console.log("Reading config from: " + argv.yamlConfig);
+
+      const fileContents = fs.readFileSync(argv.yamlConfig, "utf8");
+      const yamlParams = yaml.safeLoad(fileContents);
+
+			argv = {...yamlParams, ...argv};
+		}
+*/
     let purl;
     if (argv.url) {
       // Scope for crawl, default to the domain of the URL
@@ -79,8 +245,8 @@ class argParser {
     }
 
     for (const opt of argv.waitUntil) {
-      if (!this.constants.WAIT_UNTIL_OPTS.includes(opt)) {
-        throw new Error("Invalid waitUntil option, must be one of: " + this.constants.WAIT_UNTIL_OPTS.join(","));
+      if (!WAIT_UNTIL_OPTS.includes(opt)) {
+        throw new Error("Invalid waitUntil option, must be one of: " + WAIT_UNTIL_OPTS.join(","));
       }
     }
 
@@ -94,9 +260,9 @@ class argParser {
     }
     argv.behaviors.forEach((x) => behaviorOpts[x] = true);
     if (argv.logging.includes("behaviors")) {
-      behaviorOpts.log = this.constants.BEHAVIOR_LOG_FUNC;
+      behaviorOpts.log = BEHAVIOR_LOG_FUNC;
     } else if (argv.logging.includes("behaviors-debug")) {
-      behaviorOpts.log = this.constants.BEHAVIOR_LOG_FUNC;
+      behaviorOpts.log = BEHAVIOR_LOG_FUNC;
       this.behaviorsLogDebug = true;
     }
     this.behaviorOpts = JSON.stringify(behaviorOpts);
@@ -214,4 +380,7 @@ class argParser {
   }
 }
 
-module.exports.argParser = argParser;
+  
+module.exports.parseArgs = function(argv) {
+  return new ArgParser().parseArgs(argv);
+};
